requirements.txt
flask>=3.0.0
odfpy>=1.4.1
python-dateutil>=2.8.2
